requirements.txt
allure-python-commons
pytest
requests
jsonschema
colorlog
allure-pytest
pycryptodome
pyyaml
pytest-check
colorama
